+++ packages/client/components/DiscussionThreadInput.tsx
@@ -15,17 +15,19 @@ import convertToTaskContent from '~/utils/draftjs/convertToTaskContent'
 import isAndroid from '~/utils/draftjs/isAndroid'
 import {DiscussionThreadInput_discussion} from '~/__generated__/DiscussionThreadInput_discussion.graphql'
 import {DiscussionThreadInput_viewer} from '~/__generated__/DiscussionThreadInput_viewer.graphql'
+import {useBeforeUnload} from '../hooks/useBeforeUnload'
+import useInitialLocalState from '../hooks/useInitialLocalState'
+import CreateTaskMutation from '../mutations/CreateTaskMutation'
+import {PALETTE} from '../styles/paletteV3'
 import anonymousAvatar from '../styles/theme/images/anonymous-avatar.svg'
+import {isViewerTypingInTask} from '../utils/viewerTypingUtils'
+import AddPollButton from './AddPollButton'
+import AddTaskButton from './AddTaskButton'
 import Avatar from './Avatar/Avatar'
 import {DiscussionThreadables} from './DiscussionThreadList'
+import SendCommentButton from './SendCommentButton'
 import CommentEditor from './TaskEditor/CommentEditor'
 import {ReplyMention, SetReplyMention} from './ThreadedItem'
-import {PALETTE} from '../styles/paletteV3'
-import CreateTaskMutation from '../mutations/CreateTaskMutation'
-import AddTaskButton from './AddTaskButton'
-import SendCommentButton from './SendCommentButton'
-import {isViewerTypingInTask} from '../utils/viewerTypingUtils'
-import {useBeforeUnload} from '../hooks/useBeforeUnload'
 
 const Wrapper = styled('div')<{isReply: boolean; isDisabled: boolean}>(({isDisabled, isReply}) => ({
   display: 'flex',
@@ -56,8 +58,10 @@ const EditorWrap = styled('div')({
   margin: '14px 0'
 })
 
-const TaskContainer = styled('div')({
+const ActionsContainer = styled('div')({
   display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
   borderTop: `1px solid ${PALETTE.SLATE_200}`,
   padding: 6
 })
@@ -103,7 +107,11 @@ const DiscussionThreadInput = forwardRef((props: Props, ref: any) => {
   const [lastTypedTimestamp, setLastTypedTimestamp] = useState<Date>()
   const allowTasks = allowedThreadables.includes('task')
   const allowComments = allowedThreadables.includes('comment')
-
+  // Quick & Dirty Feature Flag
+  const allowPolls = false
+  // const allowPolls = allowedThreadables.includes('poll')
+  useInitialLocalState(discussionId, 'isAnonymousComment', false)
+  useInitialLocalState(discussionId, 'replyingToCommentId', '')
   useBeforeUnload(() => {
     EditCommentingMutation(
       atmosphere,
@@ -227,6 +235,9 @@ const DiscussionThreadInput = forwardRef((props: Props, ref: any) => {
     CreateTaskMutation(atmosphere, {newTask}, {})
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  const addPoll = () => {}
+
   useEffect(() => {
     const focusListener = () => {
       setCanCreateTask(!isViewerTypingInTask())
@@ -240,6 +251,7 @@ const DiscussionThreadInput = forwardRef((props: Props, ref: any) => {
     }
   }, [])
 
+  const isActionsContainerVisible = allowTasks || allowPolls
   const avatar = isAnonymousComment ? anonymousAvatar : picture
   return (
     <Wrapper data-cy={`${dataCy}-wrapper`} ref={ref} isReply={isReply} isDisabled={isDisabled}>
@@ -265,10 +277,15 @@ const DiscussionThreadInput = forwardRef((props: Props, ref: any) => {
           onSubmit={onSubmit}
         />
       </CommentContainer>
-      {allowTasks && (
-        <TaskContainer>
-          <AddTaskButton dataCy={`${dataCy}-add`} onClick={addTask} disabled={!canCreateTask} />
-        </TaskContainer>
+      {isActionsContainerVisible && (
+        <ActionsContainer>
+          {allowTasks && (
+            <AddTaskButton dataCy={`${dataCy}-task`} onClick={addTask} disabled={!canCreateTask} />
+          )}
+          {allowPolls && (
+            <AddPollButton dataCy={`${dataCy}-poll`} onClick={addPoll} disabled={!canCreateTask} />
+          )}
+        </ActionsContainer>
       )}
     </Wrapper>
   )