+++ packages/client/hooks/useMeetingMemberAvatars.ts
@@ -1,12 +1,12 @@
 import graphql from 'babel-plugin-relay/macro'
 import {useMemo} from 'react'
 import {readInlineData, _FragmentRefs} from 'relay-runtime'
-import {useMeetingMemberAvatars_meeting} from '../__generated__/useMeetingMemberAvatars_meeting.graphql'
+import {useMeetingMemberAvatars_meeting$key} from '../__generated__/useMeetingMemberAvatars_meeting.graphql'
 import useAtmosphere from './useAtmosphere'
 
 const useMeetingMemberAvatars = (meetingRef: _FragmentRefs<string>) => {
   const atmosphere = useAtmosphere()
-  const meeting = readInlineData<useMeetingMemberAvatars_meeting>(
+  const meeting = readInlineData<useMeetingMemberAvatars_meeting$key>(
     graphql`
       fragment useMeetingMemberAvatars_meeting on NewMeeting @inline {
         id