+++ CHANGELOG.md
@@ -5,6 +5,42 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## 6.26.0 2021-Sep-01
+
+### Fixed
+
+- Fix bad lowercase call (#5332)
+- Fix stripeSucceedPayment (#5335)
+
+### Added
+
+- Animate reflection into Spotlight (#5271)
+
+## 6.25.0 2021-Aug-25
+
+### Fixed
+
+- TeamMemberStageItems rendering unnecessarily (#5322)
+- New/Added Jira Stories are auto-archived by default (#5321)
+- Viewer object is non-null (#5310)
+
+### Changed
+
+- Upgrade to latest graphql-tools packages (#5321)
+- Upgrade to Relay v11, React v17 (#5294)
+- Server checks 'x-application-authorization' for app auth first (#5318)
+- Client always sends auth via `x-application-authorization' (#5291)
+
+### Added
+
+- Polls table in PG (#5231)
+
+## 6.24.1 2021-Aug-23
+
+### Fixed
+
+- Remove logrocket from client, it was still running even without a key
+
 ## 6.24.0 2021-Aug-19
 
 ### Added